Update Next.js, React versions, and enable experimental reactCompiler
In this commit, we have updated the versions of Next.js, React, React-DOM in both package.json and pnpm-lock.yaml. The version for Next.js is rolling up to 15.0.0-rc.0, while React and ReactDOM are updated to 19.0.0-rc.
Moreover, we have enabled the experimental reactCompiler setting inside the nextConfig object located in next.config.js.